feat: migrate app.jsx to app.tsx add mode development option for webpack
